Creacion de la Navbar y del Footer
